requirements.txt
# Claude Session Manager Dependencies
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
httpx>=0.25.0
python-dotenv>=1.0.0
mcp>=1.0.0
